requirements.txt
opencv-python>=4.9.0
Pillow>=10.0.0
